Fixed bug that was sign extenging the history field for correlation messages, giving a bogus branch history

+++ src/dqr.cpp
@@ -6416,7 +6416,7 @@ TraceDqr::DQErr SliceFileParser::parseCorrelation(NexusMessage &nm,Analytics &an
 
 		bits += width;
 
-		nm.correlation.history = (int)tmp;
+		nm.correlation.history = tmp;
 
 		break;
 	default:
